ajusta arquivos de pipe e fifo e arquivos utils
