+++ sistemas-operacionais/trabalhoFinal/utils/normalize.js
@@ -1,0 +1,7 @@
+const regex = /\s/g;
+
+const normalize = (str) => {
+    return str.toLowerCase().replace(regex, '');
+}
+
+module.exports = normalize;
