+++ sistemas-operacionais/trabalhoFinal/utils/variaveis.js
@@ -1,9 +1,3 @@
-const regex = /\s/g;
-
-const normalize = (str) => {
-    return str.toLowerCase().replace(regex, '');
-}
-
 const STATUS_ERROR = 500;
 const STATUS_NOT_FOUND = 404;
 
@@ -15,7 +9,6 @@ module.exports = {
     PORT_1: PORT_1,
     PORT_2: PORT_2,
     PORT_3: PORT_3,
-    normalize: normalize,
     STATUS_ERROR: STATUS_ERROR,
     STATUS_NOT_FOUND: STATUS_NOT_FOUND,
 };
